This version has pyfiins separated into the data simulation part, still called pyfiins.py, and a data reduction part called dataprocessing.py. Data are written to a FITS file by pyfiins and read from that file by dataprocessing.
Secondly, the interferograms are now calculated by interpolation on the uv cube, and not by an exact calculation. This is more than 100 times faster.

Lastly, the model sky can now be read from a data cube generated earlier as well as by the old method of simulating a list of sources in an Excel file.